Implement product details feature with local data source and repository pattern
- Added ProductEntity and ProductModel to represent product data.
- Created ProductLocalDataSource to fetch products from a local JSON file.
- Implemented ProductRepository to manage product data retrieval.
- Developed GetProducts use case for fetching products.
- Introduced ProductCubit for state management of product data.
- Updated UI components to display product details, including images, titles, descriptions, and action buttons.
- Enhanced navigation to pass product data to details view.
- Added new assets for product images and JSON files for local data.
- Refactored existing code for better structure and readability.

+++ lib/Feature/item_details/presentation/views/widget/custom_users_review_header.dart
@@ -2,8 +2,8 @@ import 'package:decore_app/core/utils/app_theme.dart';
 import 'package:flutter/material.dart';
 
 class CustomUserReviewsHeader extends StatelessWidget {
-  final int totalStars;
-  final int filledStars;
+  final double totalStars;
+  final double filledStars;
   final Color filledColor;
   final Color emptyColor;
   final double starSize;
@@ -12,7 +12,7 @@ class CustomUserReviewsHeader extends StatelessWidget {
     super.key,
     this.totalStars = 5,
     this.filledStars = 4,
-    this.filledColor =  AppTheme.primaryColor,
+    this.filledColor = AppTheme.primaryColor,
     this.emptyColor = AppTheme.primaryColor,
     this.starSize = 20,
   });
@@ -30,9 +30,8 @@ class CustomUserReviewsHeader extends StatelessWidget {
             fontSize: 16,
           ),
         ),
-
         Row(
-          children: List.generate(totalStars, (index) {
+          children: List.generate(totalStars.toInt(), (index) {
             return Icon(
               index < filledStars ? Icons.star : Icons.star_border,
               color: filledColor,
